[USER32] - Fix overflow check in SetScrollRange. CID 731535 CORE-12763 #resolve
svn path=/trunk/; revision=74059

+++ reactos/win32ss/user/user32/controls/scrollbar.c
@@ -1701,7 +1701,7 @@ SetScrollRange(HWND hWnd, INT nBar, INT nMinPos, INT nMaxPos, BOOL bRedraw)
   pWnd = ValidateHwnd(hWnd);
   if ( !pWnd ) return FALSE;
 
-  if ((nMaxPos - nMinPos) > MAXLONG)
+  if (((LONGLONG)nMaxPos - nMinPos) > MAXLONG)
   {
      SetLastError(ERROR_INVALID_SCROLLBAR_RANGE);
      return FALSE;
